Trim mt_evtMain 00361140 linked field

+++ src/Event/Fcl/fclShopMisc.c
@@ -14188,7 +14188,7 @@ void FUN_00401420(int param_1,int param_2)
 // FUN_004014E0 NONMATCHING
 
 
-u32 FUN_004014e0(u64 param_1)
+u32 FUN_004014e0(u32 param_1)
